updated tree-sitter grammar to support bpf map field macros
bpf map field macros at the current projects state are unsupported. To add this
a change in tree-sitter grammar was required. That is, the
_field_declaration_list_item rule now has an additional choice that checks for
a type_field_declarator_macro or a uint_field_declarator_macro which are
two new rules. With these rules a preproc_call_expression will be seen
within the tree to be parsed by a query.